Change image sampling to follow paper Move script part of train.py to a different file Add retry loop to opening video files Add periodic model checkpoints to training
